react a pod reset properly
we have a use case when one deletes a pod. now the application keeps
reporting the service associated with the pod still has anomalies
which is wrong cause the delete's purpose was to reset the state.
the pod name changes after a delete thus we can use the fact to
detect a reset. let's zero anomalies counts for the pod. the rest
stats should remain the same as we expect them have a value.